Coligado

See merge request new-dcr/dcr-backend!4

+++ src/main/java/com/dcr/api/controller/MatrizProdutoController.java
@@ -16,10 +16,17 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.dcr.api.model.as400.Cadcor;
+import com.dcr.api.model.as400.Matriitm;
 import com.dcr.api.model.as400.Matriprd;
+import com.dcr.api.model.dto.MatriitmDTO;
+import com.dcr.api.model.dto.MatriprdComCorDTO;
+import com.dcr.api.model.dto.MatriprdComCorIdDTO;
 import com.dcr.api.model.dto.MatriprdDTO;
+import com.dcr.api.model.keys.MatriitmKey;
 import com.dcr.api.response.MatriprdResponse;
 import com.dcr.api.response.ProdutoPendenciaResponse;
+import com.dcr.api.service.as400.MatriitmService;
 import com.dcr.api.service.as400.MatriprdService;
 import com.dcr.api.utils.Auxiliar;
 
@@ -35,6 +42,9 @@ public class MatrizProdutoController {
 
 	@Autowired
 	MatriprdService service;
+	
+	@Autowired
+	MatriitmService corService;
 	
 	@GetMapping(value = "/getAll", produces = "application/json")
 	@Operation(summary = "Busca todas as Matrizes de produto")
@@ -117,6 +127,75 @@ public class MatrizProdutoController {
 		}   
 	}
 	
+	@PutMapping(value = "/createComCor", produces = "application/json")
+	@Operation(summary = "Cria um tipo de produto")
+	@ApiResponses(value = {
+	        @ApiResponse(responseCode = "201", description = "Ok"),
+	        @ApiResponse(responseCode = "400", description = "Esse Matriz de produto já existe!"),
+	        @ApiResponse(responseCode = "500", description = "Error!")
+	})
+	@ResponseStatus(HttpStatus.CREATED)
+	public ResponseEntity<Object> createComCor(@RequestBody MatriprdComCorDTO dto, HttpServletRequest request) {
+	
+		try {
+	        Matriprd matriz = service.createComCor(dto, request);
+	        
+	        for (MatriitmDTO cor : dto.itens()) {
+	        	
+				corService.createComCor(cor, request, matriz.getIdmatriz());
+			}
+	        
+	        return ResponseEntity.status(HttpStatus.CREATED)
+		        	.header("Accept", "application/json")
+		            .body("OK");
+		} catch (Exception ae) {
+		    return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR) 
+		    			.header("Accept", "application/json")
+		        		.body(ae.getMessage());                
+		}   
+	}
+	
+	@PutMapping(value = "/updateComCor", produces = "application/json")
+	@Operation(summary = "Altera uma Matriz de produto")
+	@ApiResponses(value = {
+	        @ApiResponse(responseCode = "201", description = "Ok"),
+	        @ApiResponse(responseCode = "400", description = "Matriz de produto não encontrado!"),
+	        @ApiResponse(responseCode = "500", description = "Error!")
+	})
+	@ResponseStatus(HttpStatus.OK)
+	public ResponseEntity<Object> updateComCor(@RequestBody MatriprdComCorIdDTO dto, HttpServletRequest request) {
+	
+		try {
+			for (MatriitmDTO cor : dto.itens()) {
+				MatriitmKey key = new MatriitmKey();
+				key.setIdmatriz(cor.idmatriz());
+				key.setModelo(cor.modelo());
+				key.setPartnumpd(cor.partnumpd());
+				
+				Optional<Matriitm> corOg = corService.getByID(key);
+				if(!corOg.isEmpty()) {
+					corService.update(corOg.get(), cor, request);
+				}
+				
+			}
+			Optional<Matriprd> lista = service.getByID(dto.idmatriz());
+	        if (lista.isEmpty()) {
+	            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+	                    .header("Accept", "application/json")
+	                    .body("Matriz de produto não encontrada!");
+	        }
+		
+	        service.updateComCor(lista.get(), dto,  request);
+	        return ResponseEntity.status(HttpStatus.OK)
+		        	.header("Accept", "application/json")
+		            .body("OK");
+		} catch (Exception ae) {
+		    return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR) 
+		    			.header("Accept", "application/json")
+		        		.body(ae.getMessage());                
+		}   
+	}
+	
 	@PutMapping(value = "/update", produces = "application/json")
 	@Operation(summary = "Altera uma Matriz de produto")
 	@ApiResponses(value = {
@@ -159,16 +238,26 @@ public class MatrizProdutoController {
 		try {
 			
 			Optional<Matriprd> lista = service.getByID(idmatriz);
+			
+			
+			
 	        if (lista.isEmpty()) {
 	            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
 	                    .header("Accept", "application/json")
 	                    .body("Nenhuma Matriz de produto encontrada!");
 	        }
+	        if(lista.get().getOrigprd().trim().toUpperCase().equals("MANUAL")) {
+				 service.delete(lista.get());
+			        return ResponseEntity.status(HttpStatus.OK)
+				        	.header("Accept", "application/json")
+				            .body("Matriz de produto deletada com sucesso!");
+			} else {
+				 return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+		                    .header("Accept", "application/json")
+		                    .body("Produto não pode ser excluido pois foi gerado automaticamente pelo plano de produção!");
+			}
 		
-	        service.delete(lista.get());
-	        return ResponseEntity.status(HttpStatus.OK)
-		        	.header("Accept", "application/json")
-		            .body("Matriz de produto deletada com sucesso!");
+	       
 		} catch (Exception ae) {
 		    return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR) 
 		    			.header("Accept", "application/json")
